+++ src/components/discover-more-legacy/VideoPlayer.jsx
